ref.grid-class {lsmeans}	R Documentation

The ref.grid and lsmobj classes

Description

The codebase for lsmeans is now mostly in emmeans. These two classes are simple extensions of the emmGrid class defined in emmeans, and are provided as support for objects created in older versions of lsmeans. For details, see emmGrid-class.

Author(s)

Russell V. Lenth

[Package lsmeans version 2.30-2 Index]